Backend (M-3):
- AdminDealListItem/AdminDealDetailOut: sum, amount, commission_amount now Decimal
- AdminBalanceSnapshot/AdminUserBalanceOut: amount, locked, total now Decimal
- AdminWalletListItem.total_usd_estimate now Decimal
- AdminDepositOut/AdminWithdrawalOut: amount now Decimal
- AdminTreasuryBalanceOut: accrued, withdrawn, available now Decimal
- AdminTreasuryWithdrawOut.amount now Decimal
- Removed float() casts in _to_detail, _to_list_item, _to_out, _balance_row,
  treasury_overview, _withdrawal_to_out so Decimal flows directly into schema
  constructors and is serialised as JSON string by Pydantic v2.
- Audit-log payloads for deposit.mark_paid/refund, withdrawal.*, treasury.*
  now store amount as str(decimal) so JSONB keeps full Numeric precision.
- Admin deal-delete response: refunded field now str(decimal).

Frontend (M-9):
- AdminDealListItemDto/AdminDealDetailDto: sum/amount/commission_amount now string
- AdminBalanceSnapshotDto/AdminUserBalanceDto: amount/locked/total now string
- AdminWalletListItemDto.total_usd_estimate now string
- AdminDepositDto.amount/AdminWithdrawalDto.amount now string
- AdminTreasuryBalanceDto: accrued/withdrawn/available now string
- AdminTreasuryWithdrawDto.amount now string
- Added parseDecimal(value: string | number | null | undefined) helper in
  lib/format.ts; formatAmount/formatCurrency now accept string|number.
- Updated AdminDealsPage, AdminDealDetailPage, AdminWalletsPage, AdminTreasuryPage,
  AdminWithdrawalsPage, AdminDepositsPage, AdminArbitrationPage to use
  parseDecimal(x).toFixed(decimals) so JSON-string amounts render correctly.
- Public user types (WalletBalanceDto, WalletDepositDto, WalletWithdrawalDto)
  and admin input bodies (AdminWalletAdjustBody, AdminTreasuryWithdrawBody)
  intentionally left untouched — separate follow-up.

Tests:
- tests/test_medium_precision_b1.py: assertions now parse JSON strings via
  Decimal(...) to keep the M-20 regression coverage intact.
- tests/test_admin_deals.py: assertions on detail.buyer.locked etc. use
  Decimal(...) parsing for the new wire format.
- tests/test_admin_finance.py: float() coercion on response amount before
  pytest.approx comparison.
- New tests/test_medium_precision_b2.py: integration tests asserting raw
  JSON response amounts are strings for /api/admin/wallets, deposits,
  withdrawals, and treasury overview.

+++ backend/app/schemas.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import json
 from datetime import datetime
+from decimal import Decimal
 from typing import Literal
 
 from pydantic import BaseModel, field_validator
@@ -753,10 +754,10 @@ class AdminDealListItem(BaseModel):
 
     id: int
     status: str
-    sum: float
+    sum: Decimal
     currency_code: str | None
-    amount: float | None
-    commission_amount: float | None
+    amount: Decimal | None
+    commission_amount: Decimal | None
     buyer_id: int
     buyer_username: str | None
     seller_id: int
@@ -783,9 +784,9 @@ class AdminBalanceSnapshot(BaseModel):
     username: str | None
     display_name: str
     currency_code: str | None
-    amount: float
-    locked: float
-    total: float
+    amount: Decimal
+    locked: Decimal
+    total: Decimal
 
 
 class AdminDealEventItem(BaseModel):
@@ -808,10 +809,10 @@ class AdminDealDetailOut(BaseModel):
     id: int
     status: str
     description: str
-    sum: float
+    sum: Decimal
     currency_code: str | None
-    amount: float | None
-    commission_amount: float | None
+    amount: Decimal | None
+    commission_amount: Decimal | None
     pay_commission: str
     buyer: AdminBalanceSnapshot
     seller: AdminBalanceSnapshot
@@ -1087,9 +1088,9 @@ class AdminUserBalanceOut(BaseModel):
     currency_code: str
     currency_name: str
     decimals: int
-    amount: float
-    locked: float
-    total: float
+    amount: Decimal
+    locked: Decimal
+    total: Decimal
     updated_at: datetime | None
 
 
@@ -1106,7 +1107,7 @@ class AdminWalletListItem(BaseModel):
     is_banned: bool
     is_frozen: bool
     balances: list[AdminUserBalanceOut]
-    total_usd_estimate: float
+    total_usd_estimate: Decimal
 
 
 class AdminWalletListOut(BaseModel):
@@ -1168,7 +1169,7 @@ class AdminDepositOut(BaseModel):
     username: str | None
     display_name: str
     currency_code: str
-    amount: float
+    amount: Decimal
     status: str
     provider_invoice_id: str
     pay_url: str
@@ -1192,7 +1193,7 @@ class AdminWithdrawalOut(BaseModel):
     username: str | None
     display_name: str
     currency_code: str
-    amount: float
+    amount: Decimal
     address: str
     status: str
     admin_note: str
@@ -1240,9 +1241,9 @@ class AdminTreasuryBalanceOut(BaseModel):
     currency_code: str
     currency_name: str
     decimals: int
-    accrued: float
-    withdrawn: float
-    available: float
+    accrued: Decimal
+    withdrawn: Decimal
+    available: Decimal
 
 
 class AdminTreasuryOverviewOut(BaseModel):
@@ -1293,7 +1294,7 @@ class AdminTreasuryWithdrawOut(BaseModel):
     id: int
     actor_id: int
     currency_code: str
-    amount: float
+    amount: Decimal
     address: str
     status: str
     note: str
